Changing the behavior of the default working directory
It is now the directory of the bebar file, not the current directory. So if a bebar file refers to data located in ./data, then the absolute path of the data folder will be computer from the location of the bebar file, not from the location of the current directory. This can be changed using the working directory command line parameter.

+ resolve circular dependencies